// types/CityTypes.ts
export interface City {
    name:        string;
    local_names: LocalNames;
    lat:         number;
    lon:         number;
    country:     string;
}

export interface LocalNames {
    ar?:          string;
    ascii:        string;
    bg?:          string;
    ca?:          string;
    de?:          string;
    el?:          string;
    en?:          string;
    fa?:          string;
    feature_name: string;
    fi?:          string;
    fr:           string;
    gl?:          string;
    he?:          string;
    hi?:          string;
    id?:          string;
    it?:          string;
    ja?:          string;
    la?:          string;
    lt?:          string;
    pt?:          string;
    ru?:          string;
    sr?:          string;
    th?:          string;
    tr?:          string;
    vi?:          string;
    zu?:          string;
    af?:          string;
    az?:          string;
    da?:          string;
    eu?:          string;
    hr?:          string;
    hu?:          string;
    mk?:          string;
    nl?:          string;
    no?:          string;
    pl?:          string;
    ro?:          string;
    sk?:          string;
    sl?:          string;
}